rename compression strategy
